requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.7
sendgrid==5.6.0
six==1.12.0
Werkzeug==0.14.1
